# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Myweli API
  version: 0.1.0
  description: >
    The single REST API behind every Myweli client (Flutter mobile app, the
    public Next.js web, the admin console). It replaces the app's mock
    `*ServiceInterface` implementations one slice at a time (ROADMAP Phase 3).


    This file is the **contract source of truth**. Schemas mirror the Dart DTOs
    in `mobile/lib/models/` exactly, so swapping a mock for `Api<X>Service` is a
    field-for-field match. Endpoints are added as each slice ships:
      - B0  health                      (implemented)
      - B1  GET /providers, /providers/{id}   (contract locked; impl next)
      - B2  POST /auth/otp/*, /auth/refresh   (contract locked; impl next)
    Real Mobile Money, WhatsApp/SMS and FCM are out of V1 scope here.
servers:
  - url: http://localhost:8080
    description: Local dart_frog dev server
...
paths:
...
  /me:
    get:
      tags: [auth]
      summary: Read the signed-in user's own profile (B2)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The signed-in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [auth]
      summary: Update the signed-in user's own profile (B2)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, description: "Empty string clears it." }
                avatarUrl: { type: string }
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [auth]
      summary: Delete the signed-in user's own account (B2)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /appointments:
    get:
      tags: [appointments]
      summary: List the caller's appointments (B-appt)
      description: >
        Role-scoped: a **user** token returns that user's own bookings **plus
        provider-entered (manual) bookings made to the account's OTP-verified
        phone** (auto-sync, FR-APPT-008 — the match phone is resolved
        server-side, never from the request); a **provider** token returns the
        bookings of the salon it manages (the account's linked `providerId`). A
        provider account not yet linked to a Provider gets 403.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed, noShow]
      responses:
        '200':
          description: The caller's appointments (own, or the salon's for a provider)
          content:
            application/json:
              schema:
                type: object
                required: [items, page, pageSize, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Appointment' }
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [appointments]
      summary: Book an appointment — created pending, server-priced (B-appt)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerId, serviceIds, appointmentDateTime]
              properties:
                providerId: { type: string }
                serviceIds:
                  type: array
                  items: { type: string }
                appointmentDateTime: { type: string, format: date-time }
                artistId: { type: string }
                notes: { type: string }
                depositScreenshotUrl: { type: string }
              description: >
                total/deposit/balance are NOT accepted from the client — the
                server computes them from the provider's prices + deposit policy.
      responses:
        '201':
          description: The created (pending) appointment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appointment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >
            `slot_unavailable` — the requested time isn't a free slot
            (closed/past/break/already-booked/non-aligned).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /appointments/{id}:
    get:
      tags: [appointments]
      summary: One of the caller's appointments (B-appt)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: The appointment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appointment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /appointments/{id}/cancel:
    post:
      tags: [appointments]
      summary: Cancel the caller's own appointment (B-life)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: The cancelled appointment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appointment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Not in a cancellable state (already cancelled/completed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived HS256 access token (~15 min). Obtain via /auth/otp/verify;
        refresh via /auth/refresh (rotating opaque refresh token). Applied to
        protected routes (e.g. /me) as those slices land — see docs/BACKEND.md §3.1.
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed (e.g. another user's resource)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Stable machine code, e.g. otp_expired }
        message: { type: string, description: Optional human-readable detail }
...
    User:
      type: object
      required: [id, phoneNumber]
      properties:
        id: { type: string }
        phoneNumber: { type: string }
        name: { type: string }
        email: { type: string }
        avatarUrl: { type: string }
...
    Appointment:
      type: object
      description: A booking. Mirrors Appointment.toJson(). Created `pending`.
      required:
        [id, userId, providerId, serviceIds, appointmentDate, status, totalPrice, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        providerId: { type: string }
        serviceIds:
          type: array
          items: { type: string }
        artistId: { type: string, nullable: true }
        appointmentDate: { type: string, format: date-time }
        durationMinutes:
          type: integer
          description: >
            Server-computed total service duration; backs the booking
            overlap-exclusion constraint. Read-only.
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, noShow]
        totalPrice: { type: number }
        depositAmount: { type: number }
        balanceDue: { type: number }
        cancellationWindowHours: { type: integer }
        clientName: { type: string, nullable: true }
        clientPhone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        depositScreenshotUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...